feat: separate batch code into different file
