+++ core/governance/engine.go
@@ -30,10 +30,8 @@ import (
 	"code.vegaprotocol.io/vega/core/types"
 	"code.vegaprotocol.io/vega/core/validators"
 	vgcrypto "code.vegaprotocol.io/vega/libs/crypto"
-	vgerrors "code.vegaprotocol.io/vega/libs/errors"
 	"code.vegaprotocol.io/vega/libs/num"
 	"code.vegaprotocol.io/vega/logging"
-	"golang.org/x/exp/maps"
 
 	"github.com/pkg/errors"
 )
@@ -336,111 +334,6 @@ func (e *Engine) removeActiveProposalByID(ctx context.Context, id string) {
 	}
 }
 
-func (e *Engine) evaluateBatchProposals(
-	ctx context.Context, now int64,
-) (voteClosed []*VoteClosed, addToActiveProposals []*proposal) {
-	batchIDs := maps.Keys(e.activeBatchProposals)
-	sort.Strings(batchIDs)
-
-	for _, batchID := range batchIDs {
-		batchProposal := e.activeBatchProposals[batchID]
-
-		// list of failed, declined, rejected proposals
-		var batchHasRejectedProposal bool
-		var batchHasDeclinedProposal bool
-		var closedProposals []*proposal
-		for _, propType := range batchProposal.Proposals {
-			proposal := &proposal{
-				Proposal:     propType,
-				yes:          batchProposal.yes,
-				no:           batchProposal.no,
-				invalidVotes: map[string]*types.Vote{},
-			}
-
-			// check if the market for successor proposals still exists, if not, reject the proposal
-			// in case a single proposal is rejected we can reject the whole batch
-			if nm := proposal.Terms.GetNewMarket(); nm != nil && nm.Successor() != nil {
-				if _, err := e.markets.GetMarketState(proposal.ID); err != nil {
-					proposal.RejectWithErr(types.ProposalErrorInvalidSuccessorMarket, ErrParentMarketSucceededByCompeting)
-					batchHasRejectedProposal = true
-					break
-				}
-			}
-
-			// do not check parent market, the market was either rejected when the parent was succeeded
-			// or, if the parent market state is gone (ie succession window has expired), the proposal simply
-			// loses its parent market reference
-			if proposal.ShouldClose(now) {
-				proposal.Close(e.accs, e.markets)
-				if proposal.IsPassed() {
-					e.log.Debug("Proposal passed",
-						logging.ProposalID(proposal.ID),
-						logging.ProposalBatchID(batchID),
-					)
-				} else if proposal.IsDeclined() {
-					e.log.Debug("Proposal declined",
-						logging.ProposalID(proposal.ID),
-						logging.String("details", proposal.ErrorDetails),
-						logging.String("reason", proposal.Reason.String()),
-					)
-					batchHasDeclinedProposal = true
-				}
-
-				closedProposals = append(closedProposals, proposal)
-				voteClosed = append(voteClosed, e.preVoteClosedProposal(proposal))
-			}
-		}
-
-		if batchHasRejectedProposal {
-			batchProposal.State = types.ProposalStateRejected
-			batchProposal.Reason = types.ProposalErrorProposalInBatchRejected
-
-			proposalsEvents := make([]events.Event, 0, len(batchProposal.Proposals))
-			for _, proposal := range batchProposal.Proposals {
-				if proposal.IsPassed() {
-					proposal.Reject(types.ProposalErrorProposalInBatchRejected)
-				}
-
-				proposalsEvents = append(proposalsEvents, events.NewProposalEvent(ctx, *proposal))
-			}
-
-			e.broker.Send(events.NewProposalEventFromProto(ctx, batchProposal.ToProto()))
-			e.broker.SendBatch(proposalsEvents)
-
-			delete(e.activeBatchProposals, batchProposal.ID)
-			continue
-		}
-
-		if len(closedProposals) < 1 {
-			continue
-		}
-
-		proposalEvents := make([]events.Event, 0, len(closedProposals))
-		for _, proposal := range closedProposals {
-			if proposal.IsPassed() && batchHasDeclinedProposal {
-				proposal.Decline(types.ProposalErrorProposalInBatchDeclined)
-			} else if proposal.IsPassed() {
-				addToActiveProposals = append(addToActiveProposals, proposal)
-			}
-
-			proposalEvents = append(proposalEvents, events.NewProposalEvent(ctx, *proposal.Proposal))
-			proposalEvents = append(proposalEvents, newUpdatedProposalEvents(ctx, proposal)...)
-		}
-
-		batchProposal.State = types.ProposalStatePassed
-		if batchHasDeclinedProposal {
-			batchProposal.State = types.ProposalStateDeclined
-			batchProposal.Reason = types.ProposalErrorProposalInBatchDeclined
-		}
-
-		e.broker.Send(events.NewProposalEventFromProto(ctx, batchProposal.ToProto()))
-		e.broker.SendBatch(proposalEvents)
-		delete(e.activeBatchProposals, batchProposal.ID)
-	}
-
-	return
-}
-
 func (e *Engine) OnTick(ctx context.Context, t time.Time) ([]*ToEnact, []*VoteClosed) {
 	now := t.Unix()
 
@@ -545,7 +438,7 @@ func (e *Engine) OnTick(ctx context.Context, t time.Time) ([]*ToEnact, []*VoteCl
 		e.removeActiveProposalByID(ctx, id)
 	}
 
-	// add these passed proposals from batch
+	// add these passed proposals from batch to prepare them for enactment
 	e.activeProposals = append(e.activeProposals, addToActiveProposals...)
 
 	// flush here for now
@@ -565,145 +458,6 @@ func (e *Engine) getProposal(id string) (*proposal, bool) {
 	}
 
 	return p.proposal, true
-}
-
-func (e *Engine) getBatchProposal(id string) (*batchProposal, bool) {
-	bp, ok := e.activeBatchProposals[id]
-	return bp, ok
-}
-
-func (e *Engine) validateProposalFromBatch(
-	ctx context.Context,
-	p *types.Proposal,
-	batchParams, perTypeParams types.ProposalParameters,
-) (*ToSubmit, error) {
-	batchParams.MaxEnact = perTypeParams.MaxEnact
-	batchParams.MinEnact = perTypeParams.MinEnact
-
-	if proposalErr, err := e.validateOpenProposal(p, &batchParams); err != nil {
-		p.RejectWithErr(proposalErr, err)
-
-		if e.log.IsDebug() {
-			e.log.Debug("Batch proposal rejected",
-				logging.String("proposal-id", p.ID),
-				logging.String("proposal details", p.String()),
-				logging.Error(err),
-			)
-		}
-
-		return nil, err
-	}
-
-	submit, err := e.intoToSubmit(ctx, p, &enactmentTime{current: p.Terms.EnactmentTimestamp}, false)
-	if err != nil {
-		if e.log.IsDebug() {
-			e.log.Debug("Batch proposal rejected",
-				logging.String("proposal-id", p.ID),
-				logging.String("proposal details", p.String()),
-				logging.Error(err),
-			)
-		}
-		return nil, err
-	}
-
-	return submit, nil
-}
-
-func (e *Engine) SubmitBatchProposal(
-	ctx context.Context,
-	bpsub types.BatchProposalSubmission,
-	batchID, party string,
-) ([]*ToSubmit, error) {
-	if _, ok := e.getBatchProposal(batchID); ok {
-		return nil, ErrProposalIsDuplicate // state is not allowed to change externally
-	}
-
-	timeNow := e.timeService.GetTimeNow().UnixNano()
-
-	bp := &types.BatchProposal{
-		ID:               batchID,
-		Timestamp:        timeNow,
-		ClosingTimestamp: bpsub.Terms.ClosingTimestamp,
-		Party:            party,
-		State:            types.ProposalStateOpen,
-		Reference:        bpsub.Reference,
-		Rationale:        bpsub.Rationale,
-		Proposals:        make([]*types.Proposal, 0, len(bpsub.Terms.Changes)),
-	}
-
-	var proposalsEvents []events.Event
-	defer func() {
-		e.broker.Send(events.NewProposalEventFromProto(ctx, bp.ToProto()))
-
-		if len(proposalsEvents) > 0 {
-			e.broker.SendBatch(proposalsEvents)
-		}
-	}()
-
-	proposalParamsPerProposalTermType := map[types.ProposalTermsType]*types.ProposalParameters{}
-
-	for _, change := range bpsub.Terms.Changes {
-		p := &types.Proposal{
-			ID:        change.ID,
-			BatchID:   &batchID,
-			Timestamp: timeNow,
-			Party:     party,
-			State:     bp.State,
-			Reference: bp.Reference,
-			Rationale: bp.Rationale,
-			Terms: &types.ProposalTerms{
-				ClosingTimestamp:   bp.ClosingTimestamp,
-				EnactmentTimestamp: change.EnactmentTime,
-				Change:             change.Change,
-			},
-		}
-
-		params, err := e.getProposalParams(change.Change)
-		if err != nil {
-			bp.RejectWithErr(types.ProposalErrorUnknownType, err)
-			return nil, err
-		}
-
-		proposalParamsPerProposalTermType[change.Change.GetTermType()] = params
-
-		bp.SetProposalParams(params.Clone())
-		bp.Proposals = append(bp.Proposals, p)
-	}
-
-	var toSubmits []*ToSubmit
-	errs := vgerrors.NewCumulatedErrors()
-
-	for _, p := range bp.Proposals {
-		perTypeParams := proposalParamsPerProposalTermType[p.Terms.Change.GetTermType()]
-		params := bp.ProposalParameters.Clone()
-
-		submit, err := e.validateProposalFromBatch(ctx, p, params, *perTypeParams)
-		if err != nil {
-			errs.Add(err)
-			continue
-		}
-
-		toSubmits = append(toSubmits, submit)
-	}
-
-	for _, p := range bp.Proposals {
-		if !p.IsRejected() && errs.HasAny() {
-			p.Reject(types.ProposalErrorProposalInBatchRejected)
-		}
-
-		proposalsEvents = append(proposalsEvents, events.NewProposalEvent(ctx, *p))
-	}
-
-	if errs.HasAny() {
-		bp.State = types.ProposalStateRejected
-		bp.Reason = types.ProposalErrorProposalInBatchRejected
-
-		return nil, errs
-	}
-
-	e.startBatchProposal(bp)
-
-	return toSubmits, nil
 }
 
 // SubmitProposal submits new proposal to the governance engine so it can be voted on, passed and enacted.
@@ -781,27 +535,6 @@ func (e *Engine) RejectProposal(
 
 	e.rejectProposal(ctx, p, r, errorDetails)
 	e.broker.Send(events.NewProposalEvent(ctx, *p))
-	return nil
-}
-
-func (e *Engine) RejectBatchProposal(
-	ctx context.Context, proposalID string, r types.ProposalError, errorDetails error,
-) error {
-	bp, ok := e.getBatchProposal(proposalID)
-	if !ok {
-		return ErrProposalDoesNotExist
-	}
-
-	bp.RejectWithErr(r, errorDetails)
-
-	evts := make([]events.Event, 0, len(bp.Proposals))
-	for _, proposal := range bp.Proposals {
-		e.rejectProposal(ctx, proposal, r, errorDetails)
-		evts = append(evts, events.NewProposalEvent(ctx, *proposal))
-	}
-
-	e.broker.Send(events.NewProposalEventFromProto(ctx, bp.ToProto()))
-	e.broker.SendBatch(evts)
 	return nil
 }
 
@@ -895,15 +628,6 @@ func (e *Engine) intoToSubmit(ctx context.Context, p *types.Proposal, enct *enac
 	}
 
 	return tsb, nil
-}
-
-func (e *Engine) startBatchProposal(p *types.BatchProposal) {
-	e.activeBatchProposals[p.ID] = &batchProposal{
-		BatchProposal: p,
-		yes:           map[string]*types.Vote{},
-		no:            map[string]*types.Vote{},
-		invalidVotes:  map[string]*types.Vote{},
-	}
 }
 
 func (e *Engine) startProposal(p *types.Proposal) {
@@ -1157,50 +881,6 @@ func (e *Engine) addVote(ctx context.Context, cmd types.VoteSubmission, proposal
 	}
 
 	if err := proposal.AddVote(vote); err != nil {
-		return fmt.Errorf("couldn't cast the vote: %w", err)
-	}
-
-	if e.log.IsDebug() {
-		e.log.Debug("vote submission accepted",
-			logging.PartyID(party),
-			logging.String("vote", cmd.String()),
-		)
-	}
-	e.broker.Send(events.NewVoteEvent(ctx, vote))
-
-	return nil
-}
-
-func (e *Engine) addBatchVote(ctx context.Context, batchProposal *batchProposal, cmd types.VoteSubmission, party string) error {
-	validationErrs := vgerrors.NewCumulatedErrors()
-
-	for _, proposal := range batchProposal.Proposals {
-		if err := e.validateVote(cmd, proposal, batchProposal.ProposalParameters, party); err != nil {
-			validationErrs.Add(fmt.Errorf("proposal term %q has failed with: %w", proposal.Terms.Change.GetTermType(), err))
-			continue
-		}
-	}
-
-	if validationErrs.HasAny() {
-		e.log.Debug("invalid vote submission",
-			logging.PartyID(party),
-			logging.String("vote", cmd.String()),
-			logging.Error(validationErrs),
-		)
-		return validationErrs
-	}
-
-	vote := types.Vote{
-		PartyID:                     party,
-		ProposalID:                  cmd.ProposalID,
-		Value:                       cmd.Value,
-		Timestamp:                   e.timeService.GetTimeNow().UnixNano(),
-		TotalGovernanceTokenBalance: getTokensBalance(e.accs, party),
-		TotalGovernanceTokenWeight:  num.DecimalZero(),
-		TotalEquityLikeShareWeight:  num.DecimalZero(),
-	}
-
-	if err := batchProposal.AddVote(vote); err != nil {
 		return fmt.Errorf("couldn't cast the vote: %w", err)
 	}
